requirements.txt
flask==2.3.3
librosa==0.10.1
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.0
pandas==1.5.3
matplotlib==3.7.2
seaborn==0.12.2
speechrecognition==3.10.0
pydub==0.25.1
openai-whisper==20231117
torch==2.0.1
torchaudio==2.0.2
transformers==4.33.2
praat-parselmouth==0.4.3
textstat==0.7.3
nltk==3.8.1
spacy==3.6.1
webrtcvad==2.0.10
plotly==5.15.0
gunicorn==21.2.0
